+++ src/floating-select/floating-select-creatable.component.jsx
@@ -1,14 +1,17 @@
 import React from 'react';
 import Creatable from 'react-select/lib/Creatable';
-import DropdownIndicator from './dropdown-indicator.component';
+import { DropdownCloseIndicator, DropdownOpenIndicator } from './dropdown-indicator.component';
 import styles from './styles';
+import FloatingSelectBase from './floating-select-base.component';
 
-export default function FloatingSelectCreatable(props) {
-  return (
+export default class FloatingSelectCreatable extends FloatingSelectBase {
+  render = () => (
     <Creatable
-      {...props}
-      components={{ DropdownIndicator }}
+      {...this.props}
+      components={{ DropdownIndicator: this.state.isOpen ? DropdownCloseIndicator : DropdownOpenIndicator }}
       menuPlacement="auto"
+      onMenuOpen={this.handleOpenMenu}
+      onMenuClose={this.handleCloseMenu}
       styles={styles}
     />
   );