+++ src/floating-select/floating-select.component.jsx
@@ -1,14 +1,17 @@
 import React from 'react';
 import Select from 'react-select';
-import DropdownIndicator from './dropdown-indicator.component';
+import { DropdownCloseIndicator, DropdownOpenIndicator } from './dropdown-indicator.component';
 import styles from './styles';
+import FloatingSelectBase from './floating-select-base.component';
 
-export default function FloatingSelect(props) {
-  return (
+export default class FloatingSelect extends FloatingSelectBase {
+  render = () => (
     <Select
-      {...props}
-      components={{ DropdownIndicator }}
+      {...this.props}
+      components={{ DropdownIndicator: this.state.isOpen ? DropdownCloseIndicator : DropdownOpenIndicator }}
       menuPlacement="auto"
+      onMenuOpen={this.handleOpenMenu}
+      onMenuClose={this.handleCloseMenu}
       styles={styles}
     />
   );